Replace RFC doc + standalone script with a real, CI-run test
Removes docs/symexpr-shape-inference-rfc.md and the unverified
tools/torch_onnx_symexpr_consistency_check.py evidence-generator
script, replacing both with
tests/test_symexpr_kv_cache_consistency.py: an actual pytest test
following tests/test_mnn_llm_export.py's convention
(pytest.importorskip for torch/onnxruntime, runs in the normal test
matrix) instead of a doc making claims and a script nobody runs.

Covers the same three things the script's report did, now as
assertions:

- onnxsim actually removes the past_len+seq_len Shape/Gather/Add
  scaffolding a KV-cache transformer's torch.onnx.export(dynamo=True)
  produces (with a sanity check that the raw export contains the
  pattern in the first place, so the assertion isn't vacuous);
- torch.export's own ShapeEnv independently agrees the traced graph
  combines two distinct dynamic dims -- skipped, not failed, if a
  torch version's ExportedProgram/ShapeEnv internals (not a stable
  public API) don't match what this reaches into;
- the onnxsim-simplified ONNX model matches the original eager
  PyTorch module across four concrete (batch, seq_len, past_len)
  instantiations (prefill, single-token decode with a long cache,
  batch>1, empty cache).
